requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.1
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Flask==2.2.2
frozenlist==1.3.3
future==0.18.2
graphviz==0.20.1
gunicorn==20.1.0
idna==3.4
importlib-metadata==5.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
line-bot-sdk==2.3.0
MarkupSafe==2.1.1
multidict==6.0.4
numpy==1.24.0
opencv-python==4.6.0.66
pypinyin==0.47.1
python-dotenv==0.21.0
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
transitions==0.9.0
urllib3==1.26.13
Werkzeug==2.2.2
wincertstore==0.2
yarl==1.8.2
zipp==3.11.0
